+++ src/domain/models/StartListItem.kt
@@ -4,7 +4,7 @@ data class StartListItem(
 //        val id: String,
         val userId: String,
         val competitionId: String,
-        val startTime: Float,
+        val startTime: Double,
         val participantGroup: ParticipantGroup,
         val description: String
 )
